06-Popup: added warning alert window when user tries to delete a treatment, caregiver or patient

+++ src/main/java/de/hitec/nhplus/controller/treatment/AllTreatmentController.java
@@ -1,26 +1,32 @@
 package de.hitec.nhplus.controller.treatment;
 
+import java.io.IOException;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Optional;
+
 import de.hitec.nhplus.Main;
 import de.hitec.nhplus.datastorage.CaregiverDao;
 import de.hitec.nhplus.datastorage.DaoFactory;
 import de.hitec.nhplus.datastorage.PatientDao;
 import de.hitec.nhplus.datastorage.TreatmentDao;
+import de.hitec.nhplus.model.Caregiver;
+import de.hitec.nhplus.model.Patient;
+import de.hitec.nhplus.model.Treatment;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
-import javafx.scene.control.*;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.ComboBox;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
-import de.hitec.nhplus.model.Caregiver;
-import de.hitec.nhplus.model.Patient;
-import de.hitec.nhplus.model.Treatment;
-
-import java.io.IOException;
-import java.sql.SQLException;
-import java.util.ArrayList;
 
 public class AllTreatmentController {
 
@@ -50,7 +56,7 @@ public class AllTreatmentController {
 
     @FXML
     private ComboBox<String> comboBoxPatientSelection;
-    
+
     @FXML
     private ComboBox<String> comboBoxCaregiverSelection;
 
@@ -80,11 +86,12 @@ public class AllTreatmentController {
         this.columnDescription.setCellValueFactory(new PropertyValueFactory<>("description"));
         this.tableView.setItems(this.treatments);
 
-        // Disabling the button to delete treatments as long, as no treatment was selected.
+        // Disabling the button to delete treatments as long, as no treatment was
+        // selected.
         this.buttonDelete.setDisable(true);
         this.tableView.getSelectionModel().selectedItemProperty().addListener(
-                (observableValue, oldTreatment, newTreatment) ->
-                        AllTreatmentController.this.buttonDelete.setDisable(newTreatment == null));
+                (observableValue, oldTreatment, newTreatment) -> AllTreatmentController.this.buttonDelete
+                        .setDisable(newTreatment == null));
 
         this.createPatientComboBoxData();
         this.createCaregiverComboBoxData();
@@ -107,7 +114,7 @@ public class AllTreatmentController {
         try {
             patientList = (ArrayList<Patient>) dao.readAll();
             this.patientSelection.add("Patienten auswählen...");
-            for (Patient patient: patientList) {
+            for (Patient patient : patientList) {
                 this.patientSelection.add(patient.getSurname());
             }
         } catch (SQLException exception) {
@@ -120,7 +127,7 @@ public class AllTreatmentController {
         try {
             caregiverList = (ArrayList<Caregiver>) cDao.readAll();
             this.caregiverSelection.add("Pfleger auswählen...");
-            for (Caregiver caregiver: caregiverList) {
+            for (Caregiver caregiver : caregiverList) {
                 this.caregiverSelection.add(caregiver.getSurname());
             }
         } catch (SQLException exception) {
@@ -143,7 +150,7 @@ public class AllTreatmentController {
         }
 
         Patient patient = searchPatientInList(selectedPatient);
-        if (patient !=null) {
+        if (patient != null) {
             try {
                 this.treatments.addAll(this.dao.readTreatmentsByPid(patient.getPid()));
             } catch (SQLException exception) {
@@ -154,28 +161,28 @@ public class AllTreatmentController {
 
     @FXML
     public void handleCaregiverComboBox() {
-        // String selectedCaregiver = this.comboBoxCaregiverSelection.getSelectionModel().getSelectedItem();
+        // String selectedCaregiver =
+        // this.comboBoxCaregiverSelection.getSelectionModel().getSelectedItem();
         // this.treatments.clear();
         // this.dao = DaoFactory.getDaoFactory().createTreatmentDao();
 
         // if (selectedCaregiver.equals("alle")) {
-        //     try {
-        //         this.treatments.addAll(this.dao.readAll());
-        //     } catch (SQLException exception) {
-        //         exception.printStackTrace();
-        //     }
+        // try {
+        // this.treatments.addAll(this.dao.readAll());
+        // } catch (SQLException exception) {
+        // exception.printStackTrace();
+        // }
         // }
 
         // Caregiver caregiver = searchCaregiverInList(selectedCaregiver);
         // if (caregiver !=null) {
-        //     try {
-        //         this.treatments.addAll(this.dao.readTreatmentsByPid(caregiver.getCid()));
-        //     } catch (SQLException exception) {
-        //         exception.printStackTrace();
-        //     }
+        // try {
+        // this.treatments.addAll(this.dao.readTreatmentsByPid(caregiver.getCid()));
+        // } catch (SQLException exception) {
+        // exception.printStackTrace();
+        // }
         // }
     }
-
 
     private Caregiver searchCaregiverInList(String surname) {
         for (Caregiver caregiver : this.caregiverList) {
@@ -186,7 +193,6 @@ public class AllTreatmentController {
         return null;
     }
 
-
     private Patient searchPatientInList(String surname) {
         for (Patient patient : this.patientList) {
             if (patient.getSurname().equals(surname)) {
@@ -196,29 +202,39 @@ public class AllTreatmentController {
         return null;
     }
 
-    
-
     @FXML
     public void handleDelete() {
-        int index = this.tableView.getSelectionModel().getSelectedIndex();
-        Treatment t = this.treatments.remove(index);
-        TreatmentDao dao = DaoFactory.getDaoFactory().createTreatmentDao();
-        try {
-            dao.deleteById(t.getTid());
-        } catch (SQLException exception) {
-            exception.printStackTrace();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Warnung!");
+        alert.setHeaderText("Sind Sie sicher? (Behandlung löschen)");
+        alert.setContentText("""
+                Sind Sie sicher, dass Sie den Eintrag
+                aus der Behandlungstabelle löschen möchten?
+                Laut DSGVO sind Sie verpflichtet, der Aufbewahrungspflicht von 30 Jahren nachzukommen.
+
+                (Dieser Vorgang l\u00e4sst sich NICHT mehr r\u00fcckg\u00e4ngig machen.)""");
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            int index = this.tableView.getSelectionModel().getSelectedIndex();
+            Treatment t = this.treatments.remove(index);
+            TreatmentDao dao = DaoFactory.getDaoFactory().createTreatmentDao();
+            try {
+                dao.deleteById(t.getTid());
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
         }
     }
 
     @FXML
     public void handleNewTreatment() {
-        try{
+        try {
             String selectedPatient = this.comboBoxPatientSelection.getSelectionModel().getSelectedItem();
             Patient patient = searchPatientInList(selectedPatient);
             String selectedCaregiver = this.comboBoxCaregiverSelection.getSelectionModel().getSelectedItem();
             Caregiver caregiver = searchCaregiverInList(selectedCaregiver);
             newTreatmentWindow(patient, caregiver);
-        } catch (NullPointerException exception){
+        } catch (NullPointerException exception) {
             Alert alert = new Alert(Alert.AlertType.INFORMATION);
             alert.setTitle("Information");
             alert.setHeaderText("Patient für die Behandlung fehlt!");
@@ -240,7 +256,8 @@ public class AllTreatmentController {
 
     public void newTreatmentWindow(Patient patient, Caregiver caregiver) {
         try {
-            FXMLLoader loader = new FXMLLoader(Main.class.getResource("/de/hitec/nhplus/views/treatment/NewTreatmentView.fxml"));
+            FXMLLoader loader = new FXMLLoader(
+                    Main.class.getResource("/de/hitec/nhplus/views/treatment/NewTreatmentView.fxml"));
             AnchorPane pane = loader.load();
             Scene scene = new Scene(pane);
 
@@ -258,9 +275,10 @@ public class AllTreatmentController {
         }
     }
 
-    public void treatmentWindow(Treatment treatment){
+    public void treatmentWindow(Treatment treatment) {
         try {
-            FXMLLoader loader = new FXMLLoader(Main.class.getResource("/de/hitec/nhplus/views/treatment/TreatmentView.fxml"));
+            FXMLLoader loader = new FXMLLoader(
+                    Main.class.getResource("/de/hitec/nhplus/views/treatment/TreatmentView.fxml"));
             AnchorPane pane = loader.load();
             Scene scene = new Scene(pane);
